+++ src/sql/connection/__init__.py
@@ -5,6 +5,7 @@ from sql.connection.connection import (
     is_pep249_compliant,
     PLOOMBER_DOCS_LINK_STR,
     default_alias_for_engine,
+    ResultSetCollection,
 )
 
 
@@ -15,4 +16,5 @@ __all__ = [
     "is_pep249_compliant",
     "PLOOMBER_DOCS_LINK_STR",
     "default_alias_for_engine",
+    "ResultSetCollection",
 ]