+++ src/sql/connection/connection.py
@@ -81,13 +81,35 @@ def extract_module_name_from_NoSuchModuleError(e):
     return str(e).split(":")[-1].split(".")[-1]
 
 
-"""
-When there is ModuleNotFoundError or NoSuchModuleError case
-Three types of suggestions will be shown when the missing module name is:
-1. Excepted in the pre-defined map, suggest the user to install the driver pkg
-2. Closely matched to the pre-defined map, suggest the user to type correct driver name
-3. Not found in the pre-defined map, suggest user to use valid driver pkg
-"""
+class ResultSetCollection:
+    def __init__(self) -> None:
+        self._result_sets = []
+
+    def append(self, result):
+        if result in self._result_sets:
+            self._result_sets.remove(result)
+
+        self._result_sets.append(result)
+
+    def is_last(self, result):
+        # if there are no results, return True to prevent triggering
+        # a query in the database
+        if not len(self._result_sets):
+            return True
+
+        return self._result_sets[-1] is result
+
+    def close_all(self):
+        for r in self._result_sets:
+            r.close()
+
+        self._result_sets = []
+
+    def __iter__(self):
+        return iter(self._result_sets)
+
+    def __len__(self):
+        return len(self._result_sets)
 
 
 def get_missing_package_suggestion_str(e):
@@ -355,7 +377,7 @@ class AbstractConnection(abc.ABC):
         ConnectionManager.current = self
         ConnectionManager.connections[alias] = self
 
-        self._result_sets = []
+        self._result_sets = ResultSetCollection()
 
     @abc.abstractproperty
     def dialect(self):
@@ -546,7 +568,9 @@ class SQLAlchemyConnection(AbstractConnection):
             engine, self._url
         )
 
-        self._dialect = self._get_database_information()["dialect"]
+        db_info = self._get_database_information()
+        self._dialect = db_info["dialect"]
+        self._driver = db_info["driver"]
 
         autocommit = True if config is None else config.autocommit
 
@@ -577,6 +601,10 @@ class SQLAlchemyConnection(AbstractConnection):
     @property
     def dialect(self):
         return self._dialect
+
+    @property
+    def driver(self):
+        return self._driver
 
     def _connection_execute(self, query, parameters=None):
         """Call the connection execute method
@@ -653,6 +681,11 @@ class SQLAlchemyConnection(AbstractConnection):
         with_ : list, default None
             List of CTEs to use in the query
         """
+        # mssql with pyodbc does not support multiple open result sets, so we need
+        # to close them all before issuing a new query
+        if self.dialect == "mssql" and self.driver == "pyodbc":
+            self._result_sets.close_all()
+
         if with_:
             query = self._resolve_cte(query, with_)
 
@@ -773,6 +806,7 @@ class DBAPIConnection(AbstractConnection):
         _is_duckdb_native = _check_if_duckdb_dbapi_connection(connection)
 
         self._dialect = "duckdb" if _is_duckdb_native else None
+        self._driver = None
 
         # TODO: implement the dialect blacklist and add unit tests
         self._requires_manual_commit = True if config is None else config.autocommit
@@ -790,6 +824,10 @@ class DBAPIConnection(AbstractConnection):
     @property
     def dialect(self):
         return self._dialect
+
+    @property
+    def driver(self):
+        return self._driver
 
     def raw_execute(self, query, parameters=None, with_=None):
         """Run the query without any preprocessing