+++ src/sql/run/resultset.py
@@ -4,7 +4,6 @@ from functools import reduce
 from io import StringIO
 import html
 from collections.abc import Iterable
-from collections import defaultdict
 
 
 import prettytable
@@ -15,47 +14,14 @@ from sql.telemetry import telemetry
 from sql.run.table import CustomPrettyTable
 
 
-class ResultSetsManager:
-    def __init__(self) -> None:
-        self._results = defaultdict(list)
-
-    def append_to_key(self, key, result):
-        """
-        Append object to a given key, if it already exists, it doesn't add
-        a duplicate but moves it to the end
-        """
-        results_for_key = self._results[key]
-
-        if result in results_for_key:
-            results_for_key.remove(result)
-
-        results_for_key.append(result)
-
-    def is_last_for_key(self, key, result):
-        """
-        Check if the passed result is the last one for the given key,
-        returns True if there are no results for the key
-        """
-        results_for_key = self._results[key]
-
-        # if there are no results, return True to prevent triggering
-        # a query in the database
-        if not len(results_for_key):
-            return True
-
-        return results_for_key[-1] is result
-
-
 class ResultSet(ColumnGuesserMixin):
     """
     Results of a SQL query. Fetches rows lazily (only the necessary rows to show the
     preview based on the current configuration)
     """
 
-    # user to overcome drivers limitations, see @sqlaproxy for details
-    BY_CONNECTION = ResultSetsManager()
-
     def __init__(self, sqlaproxy, config, statement=None, conn=None):
+        self._closed = False
         self._config = config
         self._statement = statement
         self._sqlaproxy = sqlaproxy
@@ -85,24 +51,31 @@ class ResultSet(ColumnGuesserMixin):
 
         self._finished_init = True
 
-        # TODO: clean up, this is redundant
         if conn:
             conn._result_sets.append(self)
 
-        ResultSet.BY_CONNECTION.append_to_key(conn, self)
-
     @property
     def sqlaproxy(self):
+        conn = self._conn
+
+        # mssql with pyodbc does not support multiple open result sets, so we need
+        # to close them all. when running this, we might've already closed the results
+        # so we need to check for that and re-open the results if needed
+        if conn.dialect == "mssql" and conn.driver == "pyodbc" and self._closed:
+            self._conn._result_sets.close_all()
+            self._sqlaproxy = self._conn.raw_execute(self._statement)
+            self._sqlaproxy.fetchmany(size=len(self._results))
+            self._conn._result_sets.append(self)
+
         # there is a problem when using duckdb + sqlalchemy: duckdb-engine doesn't
         # create separate cursors, so whenever we have >1 ResultSet, the old ones
         # become outdated and fetching their results will return the results from
         # the last ResultSet. To fix this, we have to re-issue the query
-        is_last_result = ResultSet.BY_CONNECTION.is_last_for_key(self._conn, self)
+        is_last_result = self._conn._result_sets.is_last(self)
+
         is_duckdb_sqlalchemy = (
             self._dialect == "duckdb" and not self._conn.is_dbapi_connection
         )
-
-        # TODO: re-open it if closed
 
         if (
             # skip this if we're initializing the object (we're running __init__)
@@ -114,7 +87,8 @@ class ResultSet(ColumnGuesserMixin):
             self._sqlaproxy = self._conn.raw_execute(self._statement)
             self._sqlaproxy.fetchmany(size=len(self._results))
 
-            ResultSet.BY_CONNECTION.append_to_key(self._conn, self)
+            # ensure we make his result set the last one
+            self._conn._result_sets.append(self)
 
         return self._sqlaproxy
 
@@ -442,6 +416,10 @@ class ResultSet(ColumnGuesserMixin):
             pretty.set_style(_style)
 
         return pretty
+
+    def close(self):
+        self._sqlaproxy.close()
+        self._closed = True
 
 
 def unduplicate_field_names(field_names):