+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [Fix] Fix boxplot for duckdb native ([#728](https://github.com/ploomber/jupysql/issues/728))
 * [Doc] Add Redshift tutorial
 * [Feature] Adds Redshift support for `%sqlplot boxplot`
+* [Fix] Fix error when using SQL Server with pyodbc that caused queries to fail due to multiple open result sets
 * [Fix] Improves performance when converting DuckDB results to `pandas.DataFrame`
 
 ## 0.9.0 (2023-08-01)